Improved analysis report
Some improvements made to allow easier analysis of single isocenter multi-meta dose distribution.

Added a clusters identification in the imageRGB module BaseImage class (detect_clusters method). The method returns a list of center of masses identified in the image, based on a threshold value (applied to the maximum value in the image)

Added some stuff to the PDF report:
- A new page with the 2D isodoses plot
- If profiles_coord='detect' argument passed to the publish_pdf method, dose clusters are identified on the reference dose distribution. A new page is then added to the report for each dose cluster identified showing the horizontal an vertical profiles at this position.