Clusters analysis
Added analyse_clusters method, which identifies the clusters of dose on the reference dose map, based on a given threshold relative to maximum dose.

For each cluster found, interactive plots are shown to successively get median dose over an roi inside the cluster, and x/y profiles offset and width difference.

Theses values are stored in the object under self.clusters_analysis.

When publishing pdf report, if the argument plot_clusters_analysis=True, a page is added with details for each cluster.

Goal would be to automate this analysis, but not ready for now.